# Support for project modules that produce multiple JARs (with classifiers)