--- a/core/deployment/src/main/java/io/quarkus/deployment/builditem/ArchiveRootBuildItem.java
+++ b/core/deployment/src/main/java/io/quarkus/deployment/builditem/ArchiveRootBuildItem.java
@@ -13,6 +13,7 @@ import java.util.List;
 import io.quarkus.bootstrap.model.PathsCollection;
 import io.quarkus.builder.item.SimpleBuildItem;
 import io.quarkus.fs.util.ZipUtils;
+import io.quarkus.paths.PathCollection;
 
 public final class ArchiveRootBuildItem extends SimpleBuildItem {
 
@@ -29,7 +30,7 @@ public final class ArchiveRootBuildItem extends SimpleBuildItem {
             return this;
         }
 
-        public Builder addArchiveRoots(PathsCollection paths) {
+        public Builder addArchiveRoots(PathCollection paths) {
             paths.forEach(archiveRoots::add);
             return this;
         }