--- a/core/deployment/src/main/java/io/quarkus/deployment/dev/DevModeContext.java
+++ b/core/deployment/src/main/java/io/quarkus/deployment/dev/DevModeContext.java
@@ -15,9 +15,9 @@ import java.util.Optional;
 import java.util.Set;
 
 import io.quarkus.bootstrap.app.QuarkusBootstrap;
-import io.quarkus.bootstrap.model.AppArtifactKey;
-import io.quarkus.bootstrap.model.PathsCollection;
 import io.quarkus.maven.dependency.ArtifactKey;
+import io.quarkus.paths.PathCollection;
+import io.quarkus.paths.PathList;
 
 /**
  * Object that is used to pass context data from the plugin doing the invocation
@@ -27,7 +27,7 @@ import io.quarkus.maven.dependency.ArtifactKey;
  */
 public class DevModeContext implements Serializable {
 
-    public static final CompilationUnit EMPTY_COMPILATION_UNIT = new CompilationUnit(PathsCollection.of(), null, null, null);
+    public static final CompilationUnit EMPTY_COMPILATION_UNIT = new CompilationUnit(PathList.of(), null, null, null);
 
     public static final String ENABLE_PREVIEW_FLAG = "--enable-preview";
 
@@ -59,7 +59,7 @@ public class DevModeContext implements Serializable {
     private String alternateEntryPoint;
     private QuarkusBootstrap.Mode mode = QuarkusBootstrap.Mode.DEV;
     private String baseName;
-    private final Set<AppArtifactKey> localArtifacts = new HashSet<>();
+    private final Set<ArtifactKey> localArtifacts = new HashSet<>();
 
     public boolean isLocalProjectDiscovery() {
         return localProjectDiscovery;
@@ -232,7 +232,7 @@ public class DevModeContext implements Serializable {
         this.baseName = baseName;
     }
 
-    public Set<AppArtifactKey> getLocalArtifacts() {
+    public Set<ArtifactKey> getLocalArtifacts() {
         return localArtifacts;
     }
 
@@ -245,16 +245,17 @@ public class DevModeContext implements Serializable {
         private final CompilationUnit test;
 
         private final String preBuildOutputDir;
-        private final PathsCollection sourceParents;
+        private final PathCollection sourceParents;
         private final String targetDir;
 
         ModuleInfo(Builder builder) {
             this.appArtifactKey = builder.appArtifactKey;
-            this.name = builder.name;
+            this.name = builder.name == null ? builder.appArtifactKey.toGacString() : builder.name;
             this.projectDirectory = builder.projectDirectory;
             this.main = new CompilationUnit(builder.sourcePaths, builder.classesPath,
                     builder.resourcePaths,
                     builder.resourcesOutputPath);
+
             if (builder.testClassesPath != null) {
                 this.test = new CompilationUnit(builder.testSourcePaths,
                         builder.testClassesPath, builder.testResourcePaths, builder.testResourcesOutputPath);
@@ -274,7 +275,7 @@ public class DevModeContext implements Serializable {
             return projectDirectory;
         }
 
-        public PathsCollection getSourceParents() {
+        public PathCollection getSourceParents() {
             return sourceParents;
         }
 
@@ -312,18 +313,18 @@ public class DevModeContext implements Serializable {
             private ArtifactKey appArtifactKey;
             private String name;
             private String projectDirectory;
-            private PathsCollection sourcePaths = PathsCollection.of();
+            private PathCollection sourcePaths = PathList.of();
             private String classesPath;
-            private PathsCollection resourcePaths = PathsCollection.of();
+            private PathCollection resourcePaths = PathList.of();
             private String resourcesOutputPath;
 
             private String preBuildOutputDir;
-            private PathsCollection sourceParents = PathsCollection.of();
+            private PathCollection sourceParents = PathList.of();
             private String targetDir;
 
-            private PathsCollection testSourcePaths = PathsCollection.of();
+            private PathCollection testSourcePaths = PathList.of();
             private String testClassesPath;
-            private PathsCollection testResourcePaths = PathsCollection.of();
+            private PathCollection testResourcePaths = PathList.of();
             private String testResourcesOutputPath;
 
             public Builder setArtifactKey(ArtifactKey appArtifactKey) {
@@ -341,7 +342,7 @@ public class DevModeContext implements Serializable {
                 return this;
             }
 
-            public Builder setSourcePaths(PathsCollection sourcePaths) {
+            public Builder setSourcePaths(PathCollection sourcePaths) {
                 this.sourcePaths = sourcePaths;
                 return this;
             }
@@ -351,7 +352,7 @@ public class DevModeContext implements Serializable {
                 return this;
             }
 
-            public Builder setResourcePaths(PathsCollection resourcePaths) {
+            public Builder setResourcePaths(PathCollection resourcePaths) {
                 this.resourcePaths = resourcePaths;
                 return this;
             }
@@ -366,7 +367,7 @@ public class DevModeContext implements Serializable {
                 return this;
             }
 
-            public Builder setSourceParents(PathsCollection sourceParents) {
+            public Builder setSourceParents(PathCollection sourceParents) {
                 this.sourceParents = sourceParents;
                 return this;
             }
@@ -376,7 +377,7 @@ public class DevModeContext implements Serializable {
                 return this;
             }
 
-            public Builder setTestSourcePaths(PathsCollection testSourcePaths) {
+            public Builder setTestSourcePaths(PathCollection testSourcePaths) {
                 this.testSourcePaths = testSourcePaths;
                 return this;
             }
@@ -386,7 +387,7 @@ public class DevModeContext implements Serializable {
                 return this;
             }
 
-            public Builder setTestResourcePaths(PathsCollection testResourcePaths) {
+            public Builder setTestResourcePaths(PathCollection testResourcePaths) {
                 this.testResourcePaths = testResourcePaths;
                 return this;
             }
@@ -403,12 +404,12 @@ public class DevModeContext implements Serializable {
     }
 
     public static class CompilationUnit implements Serializable {
-        private PathsCollection sourcePaths;
+        private PathCollection sourcePaths;
         private final String classesPath;
-        private final PathsCollection resourcePaths;
+        private final PathCollection resourcePaths;
         private final String resourcesOutputPath;
 
-        public CompilationUnit(PathsCollection sourcePaths, String classesPath, PathsCollection resourcePaths,
+        public CompilationUnit(PathCollection sourcePaths, String classesPath, PathCollection resourcePaths,
                 String resourcesOutputPath) {
             this.sourcePaths = sourcePaths;
             this.classesPath = classesPath;
@@ -416,7 +417,7 @@ public class DevModeContext implements Serializable {
             this.resourcesOutputPath = resourcesOutputPath;
         }
 
-        public PathsCollection getSourcePaths() {
+        public PathCollection getSourcePaths() {
             return sourcePaths;
         }
 
@@ -424,7 +425,7 @@ public class DevModeContext implements Serializable {
             return classesPath;
         }
 
-        public PathsCollection getResourcePaths() {
+        public PathCollection getResourcePaths() {
             return resourcePaths;
         }