--- a/core/deployment/src/main/java/io/quarkus/deployment/steps/ApplicationIndexBuildStep.java
+++ b/core/deployment/src/main/java/io/quarkus/deployment/steps/ApplicationIndexBuildStep.java
@@ -32,7 +32,7 @@ public class ApplicationIndexBuildStep {
 
                 @Override
                 public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                    if (file.toString().endsWith(".class")) {
+                    if (file.getFileName().toString().endsWith(".class")) {
                         log.debugf("Indexing %s", file);
                         try (InputStream stream = Files.newInputStream(file)) {
                             indexer.index(stream);