--- a/independent-projects/bootstrap/app-model/src/main/java/io/quarkus/paths/OpenPathTree.java
+++ b/independent-projects/bootstrap/app-model/src/main/java/io/quarkus/paths/OpenPathTree.java
@@ -1,0 +1,8 @@
+package io.quarkus.paths;
+
+import java.io.Closeable;
+
+public interface OpenPathTree extends PathTree, Closeable {
+
+    PathTree getOriginalTree();
+}