--- a/independent-projects/bootstrap/app-model/src/main/java/io/quarkus/paths/PathVisitor.java
+++ b/independent-projects/bootstrap/app-model/src/main/java/io/quarkus/paths/PathVisitor.java
@@ -1,0 +1,6 @@
+package io.quarkus.paths;
+
+public interface PathVisitor {
+
+    void visitPath(PathVisit visit);
+}