… core/deployment/src/main/java/io/quarkus/deployment/dev/RuntimeUpdatesProcessor.java
… core/deployment/src/main/java/io/quarkus/deployment/dev/testing/TestSupport.java
… core/deployment/src/main/java/io/quarkus/deployment/index/ApplicationArchiveBuildStep.java
… core/deployment/src/main/java/io/quarkus/deployment/index/IndexingUtil.java
… core/deployment/src/main/java/io/quarkus/deployment/mutability/DevModeTask.java
… core/deployment/src/test/java/io/quarkus/runner/classloading/DirectoryClassPathElementTestCase.java
… core/deployment/src/test/java/io/quarkus/runner/classloading/MemoryClassPathElementTestCase.java
… devtools/gradle/gradle-application-plugin/src/main/java/io/quarkus/gradle/tasks/QuarkusDev.java
… devtools/gradle/gradle-application-plugin/src/main/java/io/quarkus/gradle/tasks/QuarkusGenerateCode.java
… devtools/gradle/gradle-application-plugin/src/main/java/io/quarkus/gradle/tasks/QuarkusGradleUtils.java
… devtools/gradle/gradle-model/src/main/java/io/quarkus/gradle/tooling/GradleApplicationModelBuilder.java
… devtools/maven/src/main/java/io/quarkus/maven/DevMojo.java
… devtools/maven/src/main/java/io/quarkus/maven/GenerateCodeMojo.java
… devtools/maven/src/main/java/io/quarkus/maven/QuarkusBootstrapProvider.java
… devtools/maven/src/main/java/io/quarkus/maven/QuarkusMavenWorkspaceBuilder.java
… extensions/awt/deployment/src/main/java/io/quarkus/awt/deployment/AwtProcessor.java
… extensions/kotlin/deployment/src/main/java/io/quarkus/kotlin/deployment/KotlinCompilationProvider.java
… extensions/scala/deployment/src/main/java/io/quarkus/scala/deployment/ScalaCompilationProvider.java
… extensions/smallrye-openapi/deployment/src/main/java/io/quarkus/smallrye/openapi/deployment/SmallRyeOpenApiProcessor.java
… extensions/vertx-http/deployment/src/main/java/io/quarkus/vertx/http/deployment/HttpRootPathBuildItem.java
… extensions/vertx-http/deployment/src/main/java/io/quarkus/vertx/http/deployment/VertxHttpProcessor.java
… extensions/vertx-http/deployment/src/main/java/io/quarkus/vertx/http/deployment/devmode/console/DevConsoleProcessor.java
… extensions/webjars-locator/deployment/src/test/java/io/quarkus/webjar/locator/test/WebJarLocatorRootPathTest.java
… independent-projects/bootstrap/app-model/pom.xml
… independent-projects/bootstrap/app-model/src/main/java/io/quarkus/bootstrap/model/AppArtifactKey.java
… independent-projects/bootstrap/app-model/src/main/java/io/quarkus/bootstrap/util/BootstrapUtils.java
… independent-projects/bootstrap/app-model/src/main/java/io/quarkus/bootstrap/util/PathsUtils.java
… independent-projects/bootstrap/app-model/src/main/java/io/quarkus/bootstrap/workspace/ArtifactSources.java
… independent-projects/bootstrap/app-model/src/main/java/io/quarkus/bootstrap/workspace/DefaultArtifactSources.java
… independent-projects/bootstrap/app-model/src/main/java/io/quarkus/bootstrap/workspace/DefaultSourceDir.java