This change allows to have a "view" of an artifact content located at one or more output directories applying include and exclude filters. Essentially, it adds a possibility to support project modules that produce more than one (JAR) artifact.

Currently, the assumption in dev and test modes is that if an artifact resolves to one or more output directories, all the content found in those directories will represent the artifact's content. If however there were include/exclude filters configured for the artifact, this may lead to more classes/resources being indexed and loaded, which may lead to errors. This PR is an attempt to support these cases.

There are quite a few changes in the dependency-related API, for example:
* `ResolvedDependency.getResolvedPaths()` now returns the paths that would have been resolved by Maven/Gradle (the current implementation prefers local output dirs in dev and test modes)
* `ResolvedDependency.getContentTree()` returns an instance of `PathTree` representing artifact's content (whether it's a JAR or an output directory with possible filtering applied)
* `ArtifactSources WorkspaceModule.getSources(String classifier)` allows to obtain the sources/output content layout for a given artifact (there are also helper methods like `ArtifactSources getMainSources()`, `ArtifactSources getTestSources()`)
* `PathTreeClassPathElement` was introduced and became the main impl of the `ClassPathElement`

@famod this should help your use-case, with this change your `mocks` classifier based on `test-jar` will be supported in dev mode. I would appreciate if you could give it a try and confirm it works for you. I did add a test in this PR for your use-case.

@Postremus do you mind checking this PR and perhaps see how far it throws back the performance gains you've achieved lately? :)