@aloubyansky 2021-12-18
@glefloch this should allow us to better represent the Gradle model too, although it's not yet fully leveraged.