@aloubyansky 2021-12-18
@Postremus what did you use to profile bootstrapping? I am going to look into it.