@Postremus 2021-12-18
@aloubyansky 
I used the reproducer from #21552 for profiling to retrieve the measurements.

For bootstrap profiling on windows, I use following approach:
* Start the app using mvnw clean compile quarkus:dev -Dsuspend
* Attach a profiler (e.g. in built async profiler from IDEA)
* Connect a debugger to the app. This will let it continue running.
* Wait till the app is started.
* Stop the profiler.

On Linux, I use async-profiler directly.
* Add following configuration to quarkus-maven-plugin. Make sure to point to your async-profiler installation.
```
<configuration>
                    <jvmArgs>
                        -agentpath:/mnt/c/tools/java/async-profiler/build/libasyncProfiler.so=start,event=cpu,file=startup-cpu-profile-999-SNAPSHOT.html,interval=10000,simple
                    </jvmArgs>
                </configuration>
```
* Start the app using mvnw clean compile quarkus:dev
* Once the app is fully started, stop it.
*  In the main project directory, a .html file with the flamegraph should appear