@aloubyansky 2021-12-20
Just fyi, I found how to get it to pretty much the same numbers as in 2.6.0.Final. I'll push an update once I clean my new changes up.