@aloubyansky 2021-12-22
I pushed more changes with optimizations. Here is what I collected from 10 launches of `mvn clean quarkus:dev` and subsequent forced restarts for your reproducer project @Postremus on Linux:

* 2.6.0.Final
boot: 1412 ms
restart: 268 ms

* main
boot: 1355 ms
restart: 245 ms

* this PR
boot: 1478 ms
restart: 229 ms

So the boot is still slower (especially compared to main - great job @Postremus), although not as dramatically. The restart was consistently faster by a bit, however.