@aloubyansky 2021-12-22
I'll need to fix this on Windows, probably a path handling issue
````
2021-12-22T16:26:37.5115433Z [ERROR] Failures: 
2021-12-22T16:26:37.5116613Z [ERROR]   OpenApiStaticModelsTestCase.testMultipleStaticFiles:33 1 expectation failed.
2021-12-22T16:26:37.5117923Z JSON path paths.'/openapi'.get.operationId doesn't match.
2021-12-22T16:26:37.5118749Z Expected: someOperation
2021-12-22T16:26:37.5119225Z   Actual: null
````