@gsmet 2021-12-22
I think we should get to the bottom of the slowdown because I fail to see how a change to support producing multiple jars could impact the default case.