@aloubyansky 2021-12-22
> Also we really need to compare `main without patch` and `main with patch` because @Postremus did a lot of improvements for dev mode. If we can avoid going back to 2.6 numbers, that's better :).

That's what I did above.