@famod 2021-12-22
Not looking good in my real world project, I'm afraid:
This PR breaks test execution (not in dev mode, just regular testing via surefire) in modules that depend on our `core` module that produces both the regular jar and a `test-jar` (and a `mocks` jar, but that's only included in `qdev` profile for dev mode).
I'm getting a whopping 137 ArC depoyment errors in e.g. `api` module, all "unsatisfied dependency" errors, AFAICS.
I'll need to take a closer look.

Dev mode doesn't work either in that `api` module. I getting:
```
[error]: Build step io.quarkus.agroal.deployment.AgroalProcessor#build threw an exception: io.quarkus.runtime.configuration.ConfigurationException: Unable to find a JDBC driver corresponding to the database kind 'h2' for the default datasource. Either provide a suitable JDBC driver extension, define the driver manually, or disable the JDBC datasource by adding 'quarkus.datasource.jdbc=false' to your configuration if you don't need it.
```
which is weird, because h2 should only be used for tests; `quarkus-jdbc-h2` is a project wide _test_ dependency and h2 config only exists in `application.yaml` that is contained in `core` test-jar.