@famod 2021-12-22
Btw, another sign that something is really out of whack in test execution are warnings from ArC `An injection field must be non-static and non-final - ignoring:  ...` pointing to `private final @ConfigProperty` fields in classes that are using constructor injection (generated via lombok `@RequiredArgsConstructor`).