@aloubyansky 2021-12-22
So in your case the `api` module depends on the main `core` jar and the `core` test-jar, and the `mocks` jar in the `qdev` profile? The `core` test-jar includes `META-INF/[beans.xml|jandex.idx]` and some CDI beans?