@famod 2021-12-22
> So in your case the `api` module depends on the main `core` jar and the `core` test-jar, and the `mocks` jar in the `qdev` profile?

Exactly. I can try removing the mocks jar if that helps but that `qdev` profile is not active for tests.

> The `core` test-jar includes `META-INF/[beans.xml|jandex.idx]` and some CDI beans?

Correct, and an `application.yaml` (with ordinal 256 to "win" over the respective one in main) and a few other resources like `junit-platform.properties`.