@aloubyansky 2021-12-23
@famod I found an issue related to path filtering (PathFilter serialization/deserialization was completely broken for expressions with wildcards). I adjusted the test and fixed it in this PR.