@famod 2021-12-24
@aloubyansky I had a quick look and sorry to say that, but I'm still seeing those test issues. It doesn't even find beans for classes anymore that are located in `src/main/java` of `core`. Unsurprisingly, all other modules are affacted as well (not only `api`).

FTR, this is out test-jar config:
```xml
<execution>
    <id>test-jar</id>
    <goals>
        <goal>test-jar</goal>
    </goals>
    <configuration>
        <includes>
            <include>META-INF/beans.xml</include>
            <include>META-INF/services/org.junit.jupiter.api.extension.Extension</include>
            <include>application.yaml</include>
            <include>junit-platform.properties</include>
            <include>logging.properties</include>
            <include>org/acme/foo/testlib/**</include>
            <include>org/acme/bar/register/testlib/**</include>
            <include>org/acme/bar/register/**/TestLib*</include>
        </includes>
    </configuration>
</execution>
```