@famod 2021-12-24
So far I've ruled out jandex-maven-plugin for `main` (replaced with `beans.xml`) and the sheer presence of the `mocks` execution (and usage in `qdev`).