@famod 2021-12-24
I'll have to stop here for today. I'm wondering whether the versioning schema has something to do with it:
```xml
    <parent>
        <groupId>de.tsystems.mms.nrwegbr</groupId>
        <artifactId>register-parent</artifactId>
        <version>${revision}.${changelist}${sha1}</version>
    </parent>
<!-- ... -->
        <dependency>
            <groupId>org.acme</groupId>
            <artifactId>register-core</artifactId>
            <version>${project.version}</version>
        </dependency>
<!-- ... -->
        <dependency>
            <groupId>org.acme</groupId>
            <artifactId>register-core</artifactId>
            <version>${project.version}</version>
            <type>test-jar</type>
            <scope>test</scope>
        </dependency>
```

In `register-parent`:
```xml
    <version>${revision}.${changelist}${sha1}</version>
    <properties>
         <revision>2.0.0</revision>
         <changelist>local</changelist>
         <sha1></sha1>   <!-- can be used locally as a discriminator to keep multiple clones separated -->
<!-- ... -->
```